Reorganise the extension contribution files as per other extensions
