+++ src/webhooks/webhook.js
@@ -1,4 +1,4 @@
-import { linkPullRequest, linkBranch } from "./lib/fields.js";
+import { linkPullRequest, linkBranch } from "../lib/fields.js";
 
 aha.on("webhook", async ({ headers, payload }) => {
   const event = headers.HTTP_X_GITHUB_EVENT;
